auramodule/requirements.txt
opencv-python>=4.13.0
numpy>=2.4.2
aiohttp>=3.13.3
httpx>=0.27.0
faster-whisper>=1.2.1
zeroconf>=0.148.0
python-dotenv>=1.0.0
scipy>=1.17.0
pydantic-settings>=2.4.0
dateparser>=1.2.0
